fix some issues

+++ migrations/sqls/20251026150255-products-up.sql
@@ -2,6 +2,6 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    price DECIMAL NOT NULL,
+    price FLOAT NOT NULL,
     category VARCHAR(50)
 );
